requirements.txt
accelerate==0.27.2
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astor==0.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
Django==5.0.1
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langsmith==0.0.83
llama_cpp_python==0.2.35
MarkupSafe==2.1.4
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.6.0
packaging==23.2
pandas==2.2.0
posthog==3.3.2
protobuf==4.25.2
psutil==5.9.8
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==24.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
sentencepiece==0.1.99
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.38.1
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
whitenoise==6.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
